Parsed Southern Islands metadata in OpenCL runtime to create the list of formal arguments of the kernel. Metadata needs to be parsed in the runtime (as opposed to the driver) since the way an argument pointer is interpreted in call clSetKernelArg depends on whether the argument is a value (e.g., float) or a memory object. This argument information will need to be passed somehow to the driver using new ABI calls (next step).

+++ trunk/runtime/opencl/opencl.h
@@ -156,7 +156,7 @@ typedef void *(*opencl_arch_program_create_func_t)(
 		struct opencl_program_t *parent,
 		void *arch_device,  /* Of type 'opencl_xxx_device_t' */
 		void *binary,
-		size_t length);
+		unsigned int length);
 
 /* Free an architecture-specific program. */
 typedef void (*opencl_arch_program_free_func_t)(
@@ -165,7 +165,7 @@ typedef void (*opencl_arch_program_free_func_t)(
 /* Check if a binary blob is a valid program */
 typedef cl_bool (*opencl_arch_program_valid_binary_func_t)(
 	void *binary,
-	size_t length);
+	unsigned int length);
